GetMethod

+++ EmployeeAdminPortal/Models/Entities/Employee.cs
@@ -6,6 +6,6 @@
         public required string Name { get; set; }
         public required string Email { get; set; }
         public  string? phone { get; set; }
-        public required decimal Salary { get; set; } 
+        public required double Salary { get; set; } 
     }
 }
